requirements/base.txt
click==8.1.7
gitpython==3.1.43
requests==2.10.0
beautifulsoup4==4.11.0
newspaper3k==0.2.8
goose3==3.1.19
lxml[html_clean]==5.2.2